pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "odin-secure-comm"
version = "0.1.0"
description = "AI-to-AI secure communication layer"
authors = [{name = ""}]
readme = "README.md"
requires-python = ">=3.11,<3.12"

# Core toolchain pins
dependencies = [
	"fastapi==0.110.*",
	"starlette==0.37.*",
	"pydantic==2.6.*",
	"httpx==0.27.*",
	"prometheus-client==0.20.*",
	"google-cloud-firestore>=2.15.0,<3.0.0",
	"google-cloud-secret-manager>=2.20.0,<3.0.0",
	"cryptography>=42.0.0,<43.0.0",
	"uvicorn>=0.29.0,<0.30.0",
]
